Add the project
Jeff’s Fishing Shack (JFS) is a fishing supply store operating in the suburbs, selling a range of fishing goods to amateur fishing enthusiasts. Once a customer registers, customer doesn’t have to buy anything immediately, but will automatically be sent monthly newsletters for customers who registered for that service by email

JFS wishes to create a self-serving desktop application that can be displayed in large display unit in the premises of JFS that will allow customers to register their details, sign up for the email newsletter, purchase products and email enquiries. Customers must login to purchase items from the on-line product catalogue and to modify their account details, like their address or email address. The email address is used as the login username and the password is set by the customer on registration. The password must be eight characters or more and must contain at least two non-alphabetic characters. If these conditions are not met, registration fails. As soon as a customer selects and pays for his/her items, JFS emails out a tax invoice confirming payment for the items. Payments are done by using credit cards only.

The generation and delivery of receipts is automatic through the desktop application, but the owner of JFS, Jeff, can login to check that the desktop application is running correctly. He can view a log file which shows all the transactions (purchases) for a day (which he can select). He can also run a report of all transactions by a specific customer. Apart from the items listed below, other goods stocked by the shop include fishing line (code FL), sinkers (code SI) and swivels (code SW). The fishing line is stored by length (in 100, 200 and 300m rolls), the sinkers by weight (in grams) and the swivels by size (small, medium and large)

JFS wants to have a scalable solution, with the potential for cloud storage and future stores sharing common systems and data. Therefore, Jeff has specified a Java based JavaFX solution, and SQL Server for the desktop application development.

JFS has a small group of staff, apart from Jeff. Staff can check the current transaction for a specific customer in case a customer has a query. The JFS staff must login to perform this task. Jeff, as the owner, can also perform this task. Staff usernames and passwords are entered into the system by Jeff.